requirement.txt
# ---- Core scientific stack ----
numpy==1.23.1
pillow==9.4.0
requests==2.29.0
PyYAML==6.0
typing_extensions==4.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
scipy==1.9.1
pandas==1.5.3
protobuf==3.20.3
tqdm==4.65.0
regex==2023.5.5
tabulate==0.9.0
pytz==2023.3
tzlocal==5.0.1

# ---- Vision / media ----
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
imageio==2.9.0
imageio-ffmpeg==0.4.2
albumentations==1.3.0
kornia==0.6.0

# ---- PyTorch ecosystem (install torch/torchvision separately; see notes) ----
torchmetrics==0.6.0
einops==0.3.0
timm==0.9.2

# ---- Transformers / Diffusers ----
transformers
tokenizers
huggingface_hub
diffusers
safetensors
sentencepiece
sentence-transformers

# ---- ONNX / FAISS ----
onnx==1.14.0
onnxruntime==1.14.1
faiss-cpu==1.7.4

# ---- Data / IO / misc ----
pyarrow==7.0.0
webdataset==0.2.48
rich==13.3.5
streamlit==1.12.1
tensorboard==2.13.0
yapf==0.33.0

# ---- COCO / detection stack (pip wheels exist) ----
pycocotools==2.0.6

# ---- OpenMMLab tooling (install mmcv-full via mim; see notes) ----
mmengine==0.7.3
openmim==0.3.7

# ---- From your pip section that are clearly pip-first ----
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.0
aniso8601==9.0.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==23.1.0
autofaiss==2.15.8
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.0
click==8.1.3
clip-anytorch==2.5.2
clip-benchmark==1.4.0
clip-retrieval==2.37.0
cloudpickle==2.2.1
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
cython==0.29.34
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
embedding-reader==1.5.1
exifread-nocycle==3.0.1
filelock==3.12.0
fire==0.4.0
flask==2.3.3
flask-cors==3.0.10
flask-restful==0.3.10
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.55.0
h5py==3.8.0
humanfriendly==10.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
invisible-watermark==0.1.5
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-loader==0.2
markdown==3.4.3
markdown-it-py==2.2.0
markupsafe==2.1.2
mdurl==0.1.2
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
multilingual-clip==1.0.10
networkx==3.1
nltk==3.8.1
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.14.0
open-clip-torch==2.20.0
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==23.1
pathtools==0.1.2
prometheus-client==0.17.1
promise==2.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
pydeprecate==0.3.1
pygments==2.15.1
pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
qudida==0.0.4
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
semver==3.0.0
sentence-transformers==2.2.2
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.11
smmap==5.0.0
streamlit-drawable-canvas==0.8.0
submitit==1.4.5
sympy==1.12
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
test-tube==0.7.5
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
validators==0.20.0
wandb==0.12.21
watchdog==3.0.0
wcwidth==0.2.6
werkzeug==2.3.7
yarl==1.9.2
zipp==3.15.0

